More but still incomplete implementation or ray grid tiling.
This development path will be put on indefinite hiatus.  It was started to
provide a means to apply a tomographic regularization term to a DNNROI-like
network loss function.  However, it is now understood that ray grid tiling is
fundamentally resistant to being implemented as a differentiable function.
While there are work arounds / kludges to use such a function in training a
network, this dampens my enthusiasm to continue down that particular path.

If/when having ray grid tiling on top of PyTorch becomes useful for some other reason, this development should resume.